feat(get-network-ip): fixed interface
