+++ lib/utils/get-network-ip.js
@@ -1,0 +1,17 @@
+var os = require('os');
+
+module.exports = function getNetworkIp() {
+  var interfaces = os.networkInterfaces();
+  var addresses = [];
+
+  for (var k in interfaces) {
+    for (var k2 in interfaces[k]) {
+      var address = interfaces[k][k2];
+      if (address.family === 'IPv4' && !address.internal) {
+        addresses.push(address.address);
+      }
+    }
+  }
+
+  return addresses[0];
+};
